Move filters into separate package
I got confused between the contents of this package, so I'm splitting
the filters into a new package
