+++ src/main/java/uk/gov/register/presentation/filter/ContentTypeOptionsFilter.java
@@ -1,4 +1,4 @@
-package uk.gov.register.presentation.resource;
+package uk.gov.register.presentation.filter;
 
 import com.google.common.net.HttpHeaders;
 
@@ -9,9 +9,9 @@ import javax.ws.rs.ext.Provider;
 import java.io.IOException;
 
 @Provider
-public class ContentSecurityPolicyFilter implements ContainerResponseFilter {
+public class ContentTypeOptionsFilter implements ContainerResponseFilter {
     @Override
     public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) throws IOException {
-        responseContext.getHeaders().add(HttpHeaders.CONTENT_SECURITY_POLICY, "default-src 'self'");
+        responseContext.getHeaders().add(HttpHeaders.X_CONTENT_TYPE_OPTIONS, "nosniff");
     }
 }
