+++ src/main/java/uk/gov/register/presentation/filter/ResourceAvailabilityFilter.java
@@ -1,6 +1,7 @@
-package uk.gov.register.presentation.resource;
+package uk.gov.register.presentation.filter;
 
 import uk.gov.register.presentation.config.ResourceConfiguration;
+import uk.gov.register.presentation.resource.DownloadNotAvailable;
 import uk.gov.register.presentation.view.ViewFactory;
 
 import javax.inject.Inject;
